Handling file existence

+++ restore.sh
@@ -1,5 +1,14 @@
-rm ~/.bashrc
+#!/bin/bash
+# BASHRC
+if [ -f "~/.bashrc" ]; then
+   echo "Found pre-existing bashrc. Removing it" 
+   rm ~/.bashrc
+fi
 ln bashrc ~/.bashrc
-mkdir ~/.ssh
-rm ~/.ssh/config
-ln .ssh/config ~/ssh_config
+source ~/.bashrc
+
+# SSH
+if [ -d "~/.ssh" ]; then
+   rm ~/.ssh/config
+   ln ssh_config ~/.ssh/config
+fi
